pyproject.toml
[project]
name = "ezclip"
version = "0.1.0"
description = "AI-powered video highlights extraction API using FastAPI, Firebase, R2, and Gemini"
authors = [
    {name = "juanfcarrillo",email = "juanfi444@outlook.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "firebase-admin (>=7.0.0,<8.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "google-genai (>=1.27.0,<2.0.0)",
    "types-firebase-admin (>=0.1.1,<0.2.0)",
    "boto3 (>=1.39.15,<2.0.0)",
    "moviepy (>=2.2.1,<3.0.0)",
    "yt-dlp (>=2025.7.21,<2026.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false

[[tool.poetry.packages]]
include = "app"

[tool.poetry.group.dev.dependencies]
boto3-stubs = "^1.39.15"
pytest = "^8.4.1"
celery-types = "^0.23.0"
